+++ tests/testthat/test-load.R
@@ -5,7 +5,7 @@ test_that("connections working", {
   skip_on_cran()
   
   expect_is(load_geouy("Secciones"), "sf")
-  expect_is(load_geouy("Rutas"), "sf")
+  expect_is(load_geouy("Playas"), "sf")
   expect_is(load_geouy("Centros poblados pg"), "sf")
 })